% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateREvariance.R
\name{annotateREvariance}
\alias{annotateREvariance}
\title{Calculate random effect variances from annotations and partion of heritability.}
\usage{
annotateREvariance(gwas_locus, annotation, h2partition)
}
\arguments{
\item{gwas_locus}{data frame containing GWAS summmary statistics for fine-mapping locus.
Columns SNP (variant ID, any character string format), b (effect estimate), 
se (standard error), N (sample size), Freq1 (alternate allele frequency) are required.
All SNPs are assumed to be bi-allelic.}

\item{annotation}{data frame with functional annotations including the fine mapping locus.
Must include columns CHR, BP, SNP, CM, and additional columns named for the annotation categories
(i.e. format of ENCODE annotations at https://alkesgroup.broadinstitute.org/LDSCORE/)}

\item{h2partition}{data frame with estimated heritability partition
Must include column "Coefficient" with estimates, and "Category" with names of annotation categories.
(i.e. format of ldsc output with --h2 option)}
}
\value{
data frame with all columns in gwas_locus plus RE variance Vbeta and its variance varVbeta.
 Rows represent SNPs present in both gwas_locus and annotation.
}
\description{
Calculate random effect variances from annotations and partion of heritability.
}
