% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_docs.R
\docType{data}
\name{ldsc_h2partition}
\alias{ldsc_h2partition}
\title{Estimates of annotation-level heritability partion from Locke et al GWAS of BMI}
\format{
An object of class \code{data.frame} with 53 rows and 10 columns.
}
\usage{
ldsc_h2partition
}
\description{
Estimates of annotation-level heritability partion from Locke et al GWAS of BMI
}
\keyword{datasets}
