% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_docs.R
\docType{data}
\name{locus_annotation}
\alias{locus_annotation}
\title{ENCODE project annotations for the example locus}
\format{
a data frame with 1491 rows and 57 columns
}
\usage{
locus_annotation
}
\description{
ENCODE project annotations for the example locus
}
\keyword{datasets}
