% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_docs.R
\docType{data}
\name{locus_gwas_stats}
\alias{locus_gwas_stats}
\title{GWAS summary statistics for the example locus}
\format{
An object of class \code{data.frame} with 492 rows and 12 columns.
}
\usage{
locus_gwas_stats
}
\description{
GWAS summary statistics for the example locus
}
\keyword{datasets}
