% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REestimate.R
\name{REestimate}
\alias{REestimate}
\title{Estimate best linear unbiased predictor of SNP random effects in a GWAS locus}
\usage{
REestimate(loc_summary_df, trait_h2, loc_ld_matrix = NULL, loc_ld_file = NULL)
}
\arguments{
\item{loc_summary_df}{A data frame containing summary statistics for the locus.
Columns b, se, N, Freq1, Vbeta, varVbeta are required.}

\item{loc_ld_matrix}{A numeric matrix containing linkage disequilibrium $R^2$ for all SNPs in loc_summary_df}

\item{loc_ld_file}{A file path for the VCFtools output containing LD for all SNPs in loc_summary_df.
Only used if loc_ld_matrix is NULL.}

\item{trait_h0}{A number between 0 and 1 representing total trait heritability}
}
\value{
A data frame matching loc_summary_df with columns containing AnnoRE results appended.
}
\description{
Estimate best linear unbiased predictor of SNP random effects in a GWAS locus
}
